+++ Downloads/cs451checkers/src/main/java/com/cs451checkers/checkersgame/controller/ChatController.java
@@ -1,28 +1,16 @@
 package com.cs451checkers.checkersgame.controller;
 
-import com.cs451checkers.checkersgame.model.ChatMessage;
+import com.cs451checkers.checkersgame.model.Message;
 import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.stereotype.Controller;
 
 @Controller
 public class ChatController {
 
-    @MessageMapping("/chat.sendMessage")
-    @SendTo("/topic/public")
-    public ChatMessage sendMessage(@Payload ChatMessage chatMessage) {
-        return chatMessage;
+    @MessageMapping("/hello")
+    @SendTo("/topic/greetings")
+    public Message sendMessage(Message message) {
+        return new Message("This is test content");
     }
-
-    @MessageMapping("/chat.addUser")
-    @SendTo("/topic/public")
-    public ChatMessage addUser(@Payload ChatMessage chatMessage,
-                               SimpMessageHeaderAccessor headerAccessor) {
-        // Add username in web socket session
-        headerAccessor.getSessionAttributes().put("username", chatMessage.getSender());
-        return chatMessage;
-    }
-
 }
