+++ Downloads/cs451checkers/src/main/resources/static/css/main.css
@@ -1,110 +1,217 @@
-body , html{
-    transition: 1s ease-in-out;
+html, body {
+    position: absolute;
+    top: 0;
+    left: 0;
+    height: 100%;
     width: 100%;
-    height : 100%;
-    margin : 0 0;
-    padding: 0  0;
-}
-.square{
-    float: left;
-    width: 80px;
-    height: 80px;
+    margin: 0;
+    padding: 0;
+    color: #333333;
+    font-family: "Lato", Calibri, sans-serif;
+    font-size: 16px;
+    background-color: #3A3042;
 }
 
-.white_square{
-    background-color:#F0D2B4;
+html h2, body h2 {
+    font-size: 1.3em;
 }
-.black_square{
-    background-color :#BA7A3A;
+
+html h3, body h3 {
+    font-size: 1.1em;
 }
-.clear_float{
+
+html a, body a {
+    text-decoration: none;
+    color: #333333;
+    font-weight: 700;
+}
+
+html a:hover, body a:hover {
+    text-decoration: underline;
+}
+
+html hr, body hr {
+    border: 0;
+    height: 1px;
+    background-color: #333;
+    background-image: linear-gradient(to right, #333333, #ccc);
+}
+
+html .clearfix, body .clearfix {
     clear: both;
 }
 
-.table{
+div.column {
     position: relative;
-    width: 640px;
-    height: 640px;
-    margin: 0 auto;
-}
-.score{
-    background-color: #1aaaad;
-    color: white;
-    display: none;
-    font-size: 45px;
-    font-weight: 900;
-    height:150px;
-    border-radius: 10%;
-    left: 0px;
-    letter-spacing: 1px;
-    margin : 0 auto;
-    padding-top: 30px;
-    overflow: hidden;
-    position: absolute;
-    right: 0px;
-    text-align: center;
-    top: 15%;
-    width: 200px;
-    z-index: 8;
-    font-family: Calibri, Candara, Segoe, 'Segoe UI', Optima, Arial, sans-serif;
-    -webkit-transition: 5s ease-in-out;
-    -moz-transition: 5s ease-in-out;
-    -o-transition: 5s ease-in-out;
-    transition: 5s ease-in-out;
-}
-
-.checker{
-    top:10px;
-    left:10px;
-    width: 54px;
-    height: 54px;
-    border-radius: 50%;
-    position: absolute;
-    border: 4px solid white;
-    cursor: pointer;
-
-}
-
-.white_checker{
-    background: #CC0000;
-    -webkit-transition: 0.3s ease-in-out;
-    -moz-transition: 0.3s ease-in-out;
-    -o-transition: 0.3s ease-in-out;
-    transition: 0.3s ease-in-out;
-}
-
-.black_checker{
-    background: #00001F;
-    -webkit-transition: 0.3s ease-in-out;
-    -moz-transition: 0.3s ease-in-out;
-    -o-transition: 0.3s ease-in-out;
-    transition: 0.3s ease-in-out;
-}
-
-.black_background{
-    position: fixed;
-    width: 100%;
+    float: left;
+    overflow: auto;
     height: 100%;
-    background-color: black;
-    opacity: 0.5;
-    z-index:  7;
-    display: none;
+    min-height: 100%;
+    width: 50%;
 }
 
-@media only screen and (max-width : 640px){
-    .table{
-        width: 400px;
-        height: 400px;
-    }
-
-    .square{
-        width: 50px;
-        height: 50px;
-    }
-
-    .checker{
-        width: 32px;
-        height: 32px;
+@media screen and (max-width: 1000px) {
+    div.column {
+        width: 100% !important;
+        overflow: hidden;
+        height: auto;
     }
 }
 
+div.info, div.stats {
+    width: 70%;
+    margin: 15vmin auto 0;
+    box-sizing: border-box;
+    padding: 20px 30px;
+    background-color: #F1F1FF;
+    color: #3A3042;
+    border-radius: 3px;
+    box-shadow: 1px 1px 3px #232621;
+}
+
+div.stats {
+    margin: 50px auto !important;
+}
+
+div.stats .wrapper {
+    display: flex;
+}
+
+div.stats #player1 {
+    text-align: center;
+    display: inline-block;
+    width: 50%;
+    float: left;
+    background-color: #B93848;
+    box-sizing: border-box;
+    padding: 0 15px 20px;
+    border-top-left-radius: 3px;
+    color: #232621;
+}
+
+div.stats #player1 .capturedPiece {
+    width: 2.4vmin;
+    height: 2.4vmin;
+    background-color: #232621;
+    background-size: 65%;
+    background-repeat: no-repeat;
+    background-position: center;
+    border-radius: 6vmin;
+    display: inline-block;
+    box-sizing: border-box;
+    transition: all 0.2s linear;
+    margin: 5px;
+}
+
+div.stats #player2 {
+    text-align: center;
+    display: inline-block;
+    width: 50%;
+    float: left;
+    background-color: #232621;
+    box-sizing: border-box;
+    padding: 0 15px 20px;
+    border-top-right-radius: 3px;
+    color: #B93848;
+}
+
+div.stats #player2 .capturedPiece {
+    width: 2.4vmin;
+    height: 2.4vmin;
+    background-color: #B93848;
+    background-size: 65%;
+    background-repeat: no-repeat;
+    background-position: center;
+    border-radius: 6vmin;
+    display: inline-block;
+    box-sizing: border-box;
+    transition: all 0.2s linear;
+    margin: 5px;
+}
+
+div.stats div.turn {
+    height: 10px;
+    width: 100%;
+    background: linear-gradient(to right, #BEEE62 50%, transparent 50%);
+    border-radius: 0 0 3px 3px;
+}
+
+div.stats button#cleargame {
+    display: block;
+    border: 0;
+    color: #F1F1FF;
+    border-radius: 5px;
+    cursor: pointer;
+    margin: 20px auto 10px;
+    padding: 10px;
+    background-color: #16A8C7;
+    box-shadow: 1px 1px 3px #232621;
+    transition: background-color 0.15s linear;
+}
+
+div.stats button#cleargame:hover {
+    background-color: #0c5a6b;
+}
+
+div#board {
+    position: absolute;
+    top: calc(50% - 40vmin);
+    left: calc(50% - 40vmin);
+    width: 80vmin;
+    height: 80vmin;
+    border-radius: 5px;
+    box-shadow: 1px 1px 3px #232621;
+    background-color: #B93848;
+    overflow: hidden;
+}
+
+div#board .tile {
+    width: 10vmin;
+    height: 10vmin;
+    position: absolute;
+    background-color: #232621;
+}
+
+div#board .pieces .piece {
+    position: absolute;
+    width: 8vmin;
+    height: 8vmin;
+    background-color: #e4a6ae;
+    background-size: 65%;
+    background-repeat: no-repeat;
+    background-position: center;
+    border-radius: 6vmin;
+    display: inline-block;
+    box-sizing: border-box;
+    transition: all 0.2s linear;
+    margin-top: 1vmin;
+    margin-left: 1vmin;
+    cursor: pointer;
+    border: 1vmin solid #B93848;
+}
+
+div#board .player1pieces .piece.selected {
+    box-shadow: 0 0 10px 5px #16A8C7;
+}
+
+div#board .player2pieces .piece {
+    position: absolute;
+    width: 8vmin;
+    height: 8vmin;
+    background-color: #F1F1FF;
+    background-size: 65%;
+    background-repeat: no-repeat;
+    background-position: center;
+    border-radius: 6vmin;
+    display: inline-block;
+    box-sizing: border-box;
+    transition: all 0.2s linear;
+    margin-top: 1vmin;
+    margin-left: 1vmin;
+    cursor: pointer;
+    border: 1vmin solid #8b8bff;
+}
+
+div#board .player2pieces .piece.selected {
+    box-shadow: 0 0 10px 5px #16A8C7;
+}
