+++ Downloads/cs451checkers/src/main/resources/static/index.html
@@ -1,136 +1,49 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
 <head>
-    <title>
-
-    </title>
-    <link rel="stylesheet" href="/css/main.css">
-
-</head>
-<body id="ht">
-    <div class="checker white_checker" style="display:none"> </div>
-    <div class="checker black_checker" style="display:none"> </div>
-
-
-    <div class="square" style="display: none" id ="ht"> </div>
-    <div class="black_background" id="black_background"> </div>
-
-    <div class="score" id="score">
-        <br>
-    </div>
-    <div class="table" id="table">
-
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-        <div class="checker white_checker"> </div>
-
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-        <div class="checker black_checker"> </div>
-
-
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="clear_float"> </div>
-
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="square white_square"> </div>
-        <div class="square black_square"> </div>
-        <div class="clear_float"> </div>
-
-    </div>
-
-    <script src="/js/main.js" > </script>
+    <title>CS451 Checkerts</title>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+    <link href="/css/main.css" rel="stylesheet">
+    <script src="/js/main.js"></script>
+    <script src="/js/stompImpl.js"></script>
     <script src="/js/sockjs.min.js"></script>
     <script src="/js/stomp.min.js"></script>
+</head>
+<body>
+<div class="column">
+    <button id="connect">connect</button>
+    <button id="testing">testing</button>
+    <button id="disconnect">disconnect</button>
+</div>
+<div class="column">
+    <div class="info">
+        <h1>Checkers</h1>
+        <hr>
+    </div>
+    <div class="stats">
+        <h2>Game Statistics</h2>
+        <div class="wrapper">
+            <div id="player1">
+                <h3>Player 1 (Top)</h3>
+            </div>
+            <div id="player2">
+                <h3>Player 2 (Bottom)</h3>
+            </div>
+        </div>
+        <div class="clearfix"></div>
+        <div class="turn"></div>
+        <button id="cleargame">Reset Game</button>
+    </div>
+</div>
+<div class="column">
+    <div id="board">
+        <div class="tiles"></div>
+        <div class="pieces">
+            <div class="player1pieces">
+            </div>
+            <div class="player2pieces">
+            </div>
+        </div>
+    </div>
 </body>
 </html>
