+++ Downloads/cs451checkers/src/main/resources/static/js/main.js
@@ -1,535 +1,312 @@
-
-/*=========variabile globale=========================*/
-
-
-var square_class = document.getElementsByClassName("square");
-var white_checker_class = document.getElementsByClassName("white_checker");
-var black_checker_class = document.getElementsByClassName("black_checker");
-var table = document.getElementById("table");
-var score = document.getElementById("score");
-var black_background = document.getElementById("black_background");
-var windowHeight = window.innerHeight
-    || document.documentElement.clientHeight
-    || document.body.clientHeight;  ;
-var windowWidth =  window.innerWidth
-    || document.documentElement.clientWidth
-    || document.body.clientWidth;
-var moveLength = 80 ;
-var moveDeviation = 10;
-var Dimension = 1;
-var selectedPiece,selectedPieceindex;
-var upRight,upLeft,downLeft,downRight;  // toate variantele posibile de mers pt o  dama
-var contor = 0 , gameOver = 0;
-var bigScreen = 1;
-
-var block = [];
-var w_checker = [];
-var b_checker = [];
-var the_checker ;
-var oneMove;
-var anotherMove;
-var mustAttack = false;
-var multiplier = 1 // 2 daca face saritura 1 in caz contrat
-
-var tableLimit,reverse_tableLimit ,  moveUpLeft,moveUpRight, moveDownLeft,moveDownRight , tableLimitLeft, tableLimitRight;
-
-/*================================*/
-getDimension();
-if(windowWidth > 640){
-    moveLength = 80;
-    moveDeviation = 10;
-}
-else{
-    moveLength = 50;
-    moveDeviation = 6;
-}
-
-/*================declararea claselor=========*/
-
-var square_p = function(square,index){
-    this.id = square;
-    this.ocupied = false;
-    this.pieceId = undefined;
-    this.id.onclick = function() {
-        makeMove(index);
-    }
-}
-
-var checker = function(piece,color,square) {
-    this.id = piece;
-    this.color = color;
-    this.king = false;
-    this.ocupied_square = square;
-    this.alive = true;
-    this.attack = false;
-    if(square%8){
-        this.coordX= square%8;
-        this.coordY = Math.floor(square/8) + 1 ;
-    }
-    else{
-        this.coordX = 8;
-        this.coordY = square/8 ;
-    }
-    this.id.onclick = function  () {
-        showMoves(piece);
-    }
-}
-
-checker.prototype.setCoord = function(X,Y){
-    var x = (this.coordX - 1  ) * moveLength + moveDeviation;
-    var y = (this.coordY - 1 ) * moveLength  + moveDeviation;
-    this.id.style.top = y + 'px';
-    this.id.style.left = x + 'px';
-}
-
-checker.prototype.changeCoord = function(X,Y){
-    this.coordY +=Y;
-    this.coordX += X;
-}
-
-checker.prototype.checkIfKing = function () {
-    if(this.coordY == 8 && !this.king &&this.color == "white"){
-        this.king = true;
-        this.id.style.border = "4px solid #FFFF00";
-    }
-    if(this.coordY == 1 && !this.king &&this.color == "black"){
-        this.king = true;
-        this.id.style.border = "4px solid #FFFF00";
-    }
-}
-
-/*===============Initializarea campurilor de joc =================================*/
+/*
+HEAVILY RELIES ON THIS GUY'S IMPLEMENTATION: https://github.com/codethejason/checkers
+I'm going to change names and some functionality, but it's unlikely a lot of it will change.
+I think the biggest issue will be we just have to add spring and reloadable functionality.
 
 
-for (var i = 1; i <=64; i++)
-    block[i] =new square_p(square_class[i],i);
+ */
 
-/*==================================================*/
+window.onload = function() {
+    //TODO: to pass board state around, store moves in text format like red jump 1,2 4,4 and parse it server side to send to client to handle
+    var boardstate = [
+        [0, 1, 0, 1, 0, 1, 0, 1],
+        [1, 0, 1, 0, 1, 0, 1, 0],
+        [0, 1, 0, 1, 0, 1, 0, 1],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [2, 0, 2, 0, 2, 0, 2, 0],
+        [0, 2, 0, 2, 0, 2, 0, 2],
+        [2, 0, 2, 0, 2, 0, 2, 0]
+    ];
 
+    var pieces = [];
+    var tiles = [];
+    var ActionEnum = Object.freeze({"move":456, "jump":123});
+    var PlayerTurnEnum = Object.freeze({"red":1, "black":2});
 
-/*================initializarea damelor =================================*/
+    //distance formula
+    var distance = function (x1, y1, x2, y2) {
+        return Math.sqrt(Math.pow((x1 - x2), 2) + Math.pow((y1 - y2), 2));
+    };
 
-// damele albe
-for (var i = 1; i <= 4; i++){
-    w_checker[i] = new checker(white_checker_class[i], "white", 2*i -1 );
-    w_checker[i].setCoord(0,0);
-    block[2*i - 1].ocupied = true;
-    block[2*i - 1].pieceId =w_checker[i];
-}
+    function Piece(linkedElement, position, color) {
+        //TODO: going to have to initialize pieces and tiles in nested for loop.  That way, it can have both as a
+        //comparable ij value.
+        this.position = position;
+        this.linkedElement = linkedElement;
+        this.kinged = false;
+        //TODO: enforce 0 for black, 1 for red?
+        this.owner = color;
+        this.makeKing = function () {
+            this.kinged = true;
+            //TODO: change visual
+            //TODO: call update board state
+        };
+        this.move = function (tile) {
+            this.linkedElement.removeClass('selected');
+            //TODO: check if it's a valid move.
+            //Hand a tile in, use the current tile position this piece has to check for validity
+            //If it's not a valid move, just eat it.
+            if (Board.isValidMove(tile.position[0], tile.position[1])) {
+                console.log("In valid move check");
+                if (this.owner === PlayerTurnEnum.red && this.kinged === false) {
+                    //Checks for unkinged piece moving backwards for red
+                    if (tile.position < this.position) {
+                        return false;
+                    }
+                } else if (this.owner === PlayerTurnEnum.black && this.kinged === false) {
+                    //Checks for unkinged piece moving backwards for black
+                    if (tile.position > this.position) {
+                        return false;
+                    }
+                }
 
-for (var i = 5; i <= 8; i++){
-    w_checker[i] = new checker(white_checker_class[i], "white", 2*i );
-    w_checker[i].setCoord(0,0);
-    block[2*i].ocupied = true;
-    block[2*i].pieceId = w_checker[i];
-}
+                Board.board[this.position[0]][this.position[1]] = 0;
+                Board.board[tile.position[0]][tile.position[1]] = this.owner;
 
-for (var i = 9; i <= 12; i++){
-    w_checker[i] = new checker(white_checker_class[i], "white", 2*i - 1 );
-    w_checker[i].setCoord(0,0);
-    block[2*i - 1].ocupied = true;
-    block[2*i - 1].pieceId = w_checker[i];
-}
+                //Set new position to tile position
+                this.position = [tile.position[0], tile.position[1]];
 
-//damele negre
-for (var i = 1; i <= 4; i++){
-    b_checker[i] = new checker(black_checker_class[i], "black", 56 + 2*i  );
-    b_checker[i].setCoord(0,0);
-    block[56 +  2*i ].ocupied = true;
-    block[56+  2*i ].pieceId =b_checker[i];
-}
+                this.linkedElement.css('top', Board.dictionary[this.position[0]]);
+                this.linkedElement.css('left', Board.dictionary[this.position[1]]);
+                //TODO: add board functionality to keep track of player tiles
+                //TODO: add board functionality to update which player is on which tile
 
-for (var i = 5; i <= 8; i++){
-    b_checker[i] = new checker(black_checker_class[i], "black", 40 +  2*i - 1 );
-    b_checker[i].setCoord(0,0);
-    block[ 40 + 2*i - 1].ocupied = true;
-    block[ 40 + 2*i - 1].pieceId = b_checker[i];
-}
+                if(!this.king && (this.position[0] === 0 || this.position[0] === 7 ))
+                    this.makeKing();
 
-for (var i = 9; i <= 12; i++){
-    b_checker[i] = new checker(black_checker_class[i], "black", 24 + 2*i  );
-    b_checker[i].setCoord(0,0);
-    block[24 + 2*i ].ocupied = true;
-    block[24 + 2*i ].pieceId = b_checker[i];
-}
-
-/*========================================================*/
-
-
-
-/*================SELECTIA UNEI PIESE==============*/
-the_checker = w_checker;
-
-function showMoves (piece) {
-    /* daca a fost selectat inainte o piesa stergem drumurile ei actualizand nu drumurile  Game made by Cojocaru Calin George all rights reserved piesei noi s
-    electat
-
-    */
-
-    var match = false;
-    mustAttack = false;
-    if(selectedPiece){
-        erase_roads(selectedPiece);
-    }
-    selectedPiece = piece;
-    var i,j; // retine indicele damei
-    for ( j = 1; j <= 12; j++){
-        if(the_checker[j].id == piece){
-            i = j;
-            selectedPieceindex = j;
-            match = true;
-        }
-    }
-
-    if(oneMove && !attackMoves(oneMove)){
-        changeTurns(oneMove);
-        oneMove = undefined;
-        return false;
-    }
-    if(oneMove && oneMove != the_checker[i] ){
-        return false;
-    }
-
-    if(!match) {
-        return 0 ; // daca nu a fost gasit nicio potrivire ; se intampla cand de exemplu rosu muta iar tu apasi pe negru
-    }
-
-    /*===acum in functie de culoarea lor setez marginile si miscarile damei===*/
-    if(the_checker[i].color =="white"){
-        tableLimit = 8;
-        tableLimitRight = 1;
-        tableLimitLeft = 8;
-        moveUpRight = 7;
-        moveUpLeft = 9;
-        moveDownRight = - 9;
-        moveDownLeft = -7;
-    }
-    else{
-        tableLimit = 1;
-        tableLimitRight = 8;
-        tableLimitLeft = 1;
-        moveUpRight = -7;
-        moveUpLeft = -9;
-        moveDownRight = 9;
-        moveDownLeft = 7;
-    }
-    /*===========VERIFIC DACA POT ATACA====*/
-
-
-    attackMoves(the_checker[i]); // verifica daca am vreo miscare de atac
-
-
-    /*========DACA NU POT ATACA VERIFIC DACA POT MERGE======*/
-
-    if(!mustAttack){
-        downLeft = checkMove( the_checker[i] , tableLimit , tableLimitRight , moveUpRight , downLeft);
-        downRight = checkMove( the_checker[i] , tableLimit , tableLimitLeft , moveUpLeft , downRight);
-        if(the_checker[i].king){
-            upLeft = checkMove( the_checker[i] , reverse_tableLimit , tableLimitRight , moveDownRight , upLeft);
-            upRight = checkMove( the_checker[i], reverse_tableLimit , tableLimitLeft , moveDownLeft, upRight)
-        }
-    }
-    if(downLeft || downRight || upLeft || upRight){
-        return true;
-    }
-    return false;
-
-}
-
-
-function erase_roads(piece){
-    if(downRight) block[downRight].id.style.background = "#BA7A3A";
-    if(downLeft) block[downLeft].id.style.background = "#BA7A3A";
-    if(upRight) block[upRight].id.style.background = "#BA7A3A";
-    if(upLeft) block[upLeft].id.style.background = "#BA7A3A";
-}
-
-/*=============MUTAREA PIESEI======*/
-
-function makeMove (index) {
-    var isMove = false;
-    if(!selectedPiece) // daca jocu de abea a inceput si nu a fost selectata nicio piesa
-        return false;
-    if(index != upLeft && index != upRight && index != downLeft && index != downRight){
-        erase_roads(0);
-        selectedPiece = undefined;
-        return false;
-    }
-
-    /* =========perspectiva e a jucatorului care muta ======*/
-    if(the_checker[1].color=="white"){
-        cpy_downRight = upRight;
-        cpy_downLeft = upLeft;
-        cpy_upLeft = downLeft;
-        cpy_upRight = downRight;
-    }
-    else{
-        cpy_downRight = upLeft;
-        cpy_downLeft = upRight;
-        cpy_upLeft = downRight;
-        cpy_upRight = downLeft;
-    }
-
-    if(mustAttack)  // ca sa stiu daca sar doar un rand sau 2
-        multiplier = 2;
-    else
-        multiplier = 1;
-
-
-    if(index == cpy_upRight){
-        isMove = true;
-        if(the_checker[1].color=="white"){
-            // muta piesa
-            executeMove( multiplier * 1, multiplier * 1, multiplier * 9 );
-            //elimina piesa daca a fost executata o saritura
-            if(mustAttack) eliminateCheck(index - 9);
-        }
-        else{
-            executeMove( multiplier * 1, multiplier * -1, multiplier * -7);
-            if(mustAttack) eliminateCheck( index + 7 );
-        }
-    }
-
-    if(index == cpy_upLeft){
-
-        isMove = true;
-        if(the_checker[1].color=="white"){
-            executeMove( multiplier * -1, multiplier * 1, multiplier * 7);
-            if(mustAttack)	eliminateCheck(index - 7 );
-        }
-        else{
-            executeMove( multiplier * -1, multiplier * -1, multiplier * -9);
-            if (mustAttack) eliminateCheck( index + 9 );
-        }
-    }
-
-    if(the_checker[selectedPieceindex].king){
-
-        if(index == cpy_downRight){
-            isMove = true;
-            if(the_checker[1].color=="white"){
-                executeMove( multiplier * 1, multiplier * -1, multiplier * -7);
-                if(mustAttack) eliminateCheck ( index  + 7) ;
+                Board.switchTurn();
+                Board.redraw();
+                //TODO: use websocket to send spring whatever new information via board object
+                return true;
+            } else {
+                console.log("Not a valid move");
+                //TODO: eat it
+                return false;
             }
-            else{
-                executeMove( multiplier * 1, multiplier * 1, multiplier * 9);
-                if(mustAttack) eliminateCheck ( index  - 9) ;
+        };
+
+        //tests if piece can jump anywhere
+        this.canJumpAgain = function () {
+            return (this.canJump([this.position[0] + 2, this.position[1] + 2]) ||
+                this.canJump([this.position[0] + 2, this.position[1] - 2]) ||
+                this.canJump([this.position[0] - 2, this.position[1] + 2]) ||
+                this.canJump([this.position[0] - 2, this.position[1] - 2]));
+        };
+
+        this.canJump = function (newTile) {
+            var dx = newTile[1] - this.position[1];
+            var dy = newTile[0] - this.position[0];
+            if (this.owner === PlayerTurnEnum.red && this.kinged === false) {
+                //Checks for unkinged piece moving backwards for black
+                if (newTile[0] < this.position[0]) {
+                    console.log("Failed to jump");
+                    return false;
+                }
+            } else if (this.owner === PlayerTurnEnum.black && this.kinged === false) {
+                //Checks for unkinged piece moving backwards for red
+                if (newTile[0] > this.position[0]) {
+                    console.log("Failed to jump");
+                    return false;
+                }
+            }
+
+            if (newTile[0] > 7 || newTile[1] > 7
+                || newTile[0] < 0 || newTile[1] < 0) {
+                console.log("Failed to jump 2");
+                //out of bounds check
+                return false;
+            }
+
+            //Get piece in between
+            var tileToCheckx = this.position[1] + dx / 2;
+            var tileToChecky = this.position[0] + dy / 2;
+            if (!Board.isValidMove(tileToChecky, tileToCheckx) && Board.isValidMove(newTile[0], newTile[1])) {
+                //return true;
+                for (pieceIndex in pieces) {
+                    if (pieces[pieceIndex].position[0] === tileToChecky && pieces[pieceIndex].position[1] === tileToCheckx) {
+                        if (this.owner !== pieces[pieceIndex].owner) {
+                            console.log("not sure");
+                            return pieces[pieceIndex];
+                        }
+                    }
+                }
+            } else {
+                console.log("Failed to jump 3");
+                return false;
+            }
+        };
+        this.jump = function (tile) {
+            //TODO: calculate piece to be removed
+            //if there is one, kill it, else false.
+            var canJump = this.canJump(tile.position);
+            //Get piece in between
+            if (canJump) {
+                console.log("can jump");
+                canJump.remove();
+                return true;
+            } else {
+                console.log("can't jump");
+                return false;
+            }
+
+        };
+        //TODO: add check for jump capability
+        this.remove = function () {
+            //TODO: set css to removed, add this to player's conquered pieces
+            //TODO: update global board state to remove this piece
+            //TODO: reset piece position
+            Board.redraw();
+            //TODO: use websocket to send spring whatever new information via board object
+            this.linkedElement.css("display", "none");
+            Board.board[this.position[0]][this.position[1]] = 0;
+            //reset position so it doesn't get picked up by the for loop in the canOpponentJump method
+            this.position = [];
+            //TODO: determine winner
+        };
+    }
+
+    function Tile(linkedElement, position) {
+        this.element = linkedElement;
+        this.position = position;
+        //This shit is absolute magic.
+        this.validTile = function (piece) {
+            if (distance(this.position[0], this.position[1], piece.position[0], piece.position[1]) === Math.sqrt(2)) {
+                return ActionEnum.move;
+            } else if (distance(this.position[0], this.position[1], piece.position[0], piece.position[1]) === 2 * Math.sqrt(2)) {
+                return ActionEnum.jump;
+            }
+        };
+    }
+
+    var Board = {
+        board: boardstate,
+        playerTurn: PlayerTurnEnum.red,
+        tilesElement: $('div.tiles'),
+        //TODO: potential fucking magic?
+        //dictionary to convert position in Board.board to the viewport units
+        dictionary: ["0vmin", "10vmin", "20vmin", "30vmin", "40vmin", "50vmin", "60vmin", "70vmin", "80vmin", "90vmin"],
+        //initialize board
+        initalize: function () {
+            var pieceCount = 0;
+            var tileCount = 0;
+            for (row in this.board) { //row is the index
+                for (column in this.board[row]) { //column is the index
+                    //whole set of if statements control where the tiles and pieces should be placed on the board
+                    if(row % 2 === 1) {
+                        if(column % 2 === 0) {
+                            this.tilesElement.append("<div class='tile' id='tile" + tileCount + "' style='top:" + this.dictionary[row] + ";left:" + this.dictionary[column] + ";'></div>");
+                            tiles[tileCount] = new Tile($("#tile"+tileCount), [parseInt(row), parseInt(column)]);
+                            tileCount += 1;
+                        }
+                    } else {
+                        if(column % 2 === 1) {
+                            this.tilesElement.append("<div class='tile' id='tile" + tileCount + "' style='top:" + this.dictionary[row] + ";left:"+this.dictionary[column]+  ";'></div>");
+                            tiles[tileCount] = new Tile($("#tile"+tileCount), [parseInt(row), parseInt(column)]);
+                            tileCount += 1;
+                        }
+                    }
+                    if(this.board[row][column] === 1) {
+                        $('.player1pieces').append("<div class='piece' id='" + pieceCount + "' style='top:"+this.dictionary[row]+";left:"+this.dictionary[column]+";'></div>");
+                        pieces[pieceCount] = new Piece($("#" + pieceCount), [parseInt(row), parseInt(column)], PlayerTurnEnum.red);
+                        pieceCount += 1;
+                    } else if(this.board[row][column] === 2) {
+                        $('.player2pieces').append("<div class='piece' id='" + pieceCount+ "' style='top:"+this.dictionary[row]+";left:"+this.dictionary[column]+";'></div>");
+                        pieces[pieceCount] = new Piece($("#" + pieceCount), [parseInt(row), parseInt(column)], PlayerTurnEnum.black);
+                        pieceCount += 1;
+                    }
+                }
+            }
+        },
+        //check if the location has an object
+        isValidMove: function (row, column) {
+            console.log(row); console.log(column); console.log(this.board);
+            return this.board[row][column] === 0;
+        },
+        //change the active player - also changes div.turn's CSS
+        switchTurn: function () {
+            if(this.playerTurn === PlayerTurnEnum.red) {
+                this.playerTurn = PlayerTurnEnum.black;
+                //TODO: are these ones necessary?
+                $('.turn').css("background", "linear-gradient(to right, transparent 50%, #BEEE62 50%)");
+            } else if(this.playerTurn === PlayerTurnEnum.black) {
+                this.playerTurn = PlayerTurnEnum.red;
+                $('.turn').css("background", "linear-gradient(to right, #BEEE62 50%, transparent 50%)");
+            }
+        },
+        //reset the game
+        clear: function () {
+            //TODO: resets the game board???? toggleable via switch?
+            location.reload();
+        },
+        redraw: function () {
+            console.log("Yeah does nothing");
+        }
+    };
+
+    //initialize the board
+    Board.initalize();
+
+    /***
+     Events
+     ***/
+
+    //select the piece on click if it is the player's turn
+    $('.piece').on("click", function () {
+        var selected;
+        var isPlayersTurn = ($(this).parent().attr("class").split(' ')[0] === "player" + Board.playerTurn + "pieces");
+        if(isPlayersTurn) {
+            if($(this).hasClass('selected')) {
+                selected = true;
+            }
+            $('.piece').each(function(index) {$('.piece').eq(index).removeClass('selected')});
+            if(!selected) {
+                $(this).addClass('selected');
             }
         }
+    });
 
-        if(index == cpy_downLeft){
-            isMove = true;
-            if(the_checker[1].color=="white"){
-                executeMove( multiplier * -1, multiplier * -1, multiplier * -9);
-                if(mustAttack) eliminateCheck ( index  + 9) ;
+    //reset game when clear button is pressed
+    //TODO: yeah sure, but we're going to have to fix the menu location and stuff.
+    $('#cleargame').on("click", function () {
+        Board.clear();
+    });
+
+    //move piece when tile is clicked
+    $('.tile').on("click", function () {
+        //make sure a piece is selected
+        if($('.selected').length !== 0) {
+            //find the tile object being clicked
+            var tileID = $(this).attr("id").replace(/tile/, '');
+            var tile = tiles[tileID];
+            //find the piece being selected
+            var piece = pieces[$('.selected').attr("id")];
+            //check if the tile is in range from the object
+            var inRange = tile.validTile(piece);
+            if(inRange) {
+                //if the move needed is jump, then move it but also check if another move can be made (double and triple jumps)
+                if(inRange === ActionEnum.jump) {
+                    if(piece.jump(tile)) {
+                        piece.move(tile);
+                        if (piece.canJumpAgain()) {
+                            Board.switchTurn();
+                            piece.linkedElement.addClass('selected');
+                        }
+                        //TODO: this logic is garbage, change it to be that if a player jumps, instead check for more jumps.
+                        //TODO: if there are, don't do anything and let them click, and give option to move, otherwise end turn
+                        /*if(piece.canJumpAny()) {
+                            Board.switchTurn(); //change back to original since another turn can be made
+                            piece.element.addClass('selected');
+                        }*/
+                    }
+                    //if it's regular then move it if no jumping is available
+                } else if(inRange === ActionEnum.move) {
+                    console.log("move");
+                    piece.move(tile);
+                }
+            } else {
+                console.log("fuck not in range");
             }
-            else{
-                executeMove( multiplier * -1, multiplier * 1, multiplier * 7);
-                if(mustAttack) eliminateCheck ( index  - 7) ;
-            }
         }
-    }
-
-    erase_roads(0);
-    the_checker[selectedPieceindex].checkIfKing();
-
-    if (isMove) {
-        anotherMove = undefined;
-        if(mustAttack) {
-            anotherMove = attackMoves(the_checker[selectedPieceindex]);
-        }
-        if (anotherMove){
-            oneMove = the_checker[selectedPieceindex];
-            showMoves(oneMove);
-        }
-        else{
-            oneMove = undefined;
-            changeTurns(the_checker[1]);
-            gameOver = checkIfLost();
-            if(gameOver) { setTimeout( declareWinner(),3000 ); return false};
-            gameOver = checkForMoves();
-            if(gameOver) { setTimeout( declareWinner() ,3000) ; return false};
-        }
-    }
-}
-
-/*===========MUTAREA PIESEI-SCHIMBAREA COORDONATELOR======*/
-
-function executeMove (X,Y,nSquare){
-    // schimb coordonate piesei mutate
-    the_checker[selectedPieceindex].changeCoord(X,Y);
-    the_checker[selectedPieceindex].setCoord(0,0);
-    // eliberez campul pe care il ocupa piesa si il ocup pe cel pe care il va ocupa
-    block[the_checker[selectedPieceindex].ocupied_square].ocupied = false;
-    //alert (the_checker[selectedPieceindex].ocupied_square);
-    block[the_checker[selectedPieceindex].ocupied_square + nSquare].ocupied = true;
-    block[the_checker[selectedPieceindex].ocupied_square + nSquare].pieceId = 	block[the_checker[selectedPieceindex].ocupied_square ].pieceId;
-    block[the_checker[selectedPieceindex].ocupied_square ].pieceId = undefined;
-    the_checker[selectedPieceindex].ocupied_square += nSquare;
-
-}
-
-function checkMove(Apiece,tLimit,tLimit_Side,moveDirection,theDirection){
-    if(Apiece.coordY != tLimit){
-        if(Apiece.coordX != tLimit_Side && !block[ Apiece.ocupied_square + moveDirection ].ocupied){
-            block[ Apiece.ocupied_square + moveDirection ].id.style.background = "#704923";
-            theDirection = Apiece.ocupied_square + moveDirection;
-        }
-        else
-            theDirection = undefined;
-    }
-    else
-        theDirection = undefined;
-    return theDirection;
-}
-
-
-
-function  checkAttack( check , X, Y , negX , negY, squareMove, direction){
-    if(check.coordX * negX >= 	X * negX && check.coordY *negY <= Y * negY && block[check.ocupied_square + squareMove ].ocupied && block[check.ocupied_square + squareMove].pieceId.color != check.color && !block[check.ocupied_square + squareMove * 2 ].ocupied){
-        mustAttack = true;
-        direction = check.ocupied_square +  squareMove*2 ;
-        block[direction].id.style.background = "#704923";
-        return direction ;
-    }
-    else
-        direction =  undefined;
-    return direction;
-}
-
-function eliminateCheck(indexx){
-    if(indexx < 1 || indexx > 64)
-        return  0;
-
-    var x =block[ indexx ].pieceId ;
-    x.alive =false;
-    block[ indexx ].ocupied = false;
-    x.id.style.display  = "none";
-}
-
-
-function attackMoves(ckc){
-
-    upRight = undefined;
-    upLeft = undefined;
-    downRight = undefined;
-    downLeft = undefined;
-
-    if(ckc.king ){
-        if(ckc.color == "white"){
-            upRight = checkAttack( ckc , 6, 3 , -1 , -1 , -7, upRight );
-            upLeft = checkAttack( ckc, 3 , 3 , 1 , -1 , -9 , upLeft );
-        }
-        else{
-            downLeft = checkAttack( ckc , 3, 6, 1 , 1 , 7 , downLeft );
-            downRight = checkAttack( ckc , 6 , 6 , -1, 1 ,9 , downRight );
-        }
-    }
-    if(ckc.color == "white"){
-        downLeft = checkAttack( ckc , 3, 6, 1 , 1 , 7 , downLeft );
-        downRight = checkAttack( ckc , 6 , 6 , -1, 1 ,9 , downRight );
-    }
-    else{
-        upRight = checkAttack( ckc , 6, 3 , -1 , -1 , -7, upRight );
-        upLeft = checkAttack( ckc, 3 , 3 , 1 , -1 , -9 , upLeft );
-    }
-
-    if(ckc.color== "black" && (upRight || upLeft || downLeft || downRight ) ) {
-        var p = upLeft;
-        upLeft = downLeft;
-        downLeft = p;
-
-        p = upRight;
-        upRight = downRight;
-        downRight = p;
-
-        p = downLeft ;
-        downLeft = downRight;
-        downRight = p;
-
-        p = upRight ;
-        upRight = upLeft;
-        upLeft = p;
-    }
-    if(upLeft != undefined || upRight != undefined || downRight != undefined || downLeft != undefined){
-        return true;
-
-    }
-    return false;
-}
-
-function changeTurns(ckc){
-    if(ckc.color=="white")
-        the_checker = b_checker;
-    else
-        the_checker = w_checker;
-}
-
-function checkIfLost(){
-    var i;
-    for(i = 1 ; i <= 12; i++)
-        if(the_checker[i].alive)
-            return false;
-    return true;
-}
-
-function  checkForMoves(){
-    var i ;
-    for(i = 1 ; i <= 12; i++)
-        if(the_checker[i].alive && showMoves(the_checker[i].id)){
-            erase_roads(0);
-            return false;
-        }
-    return true;
-}
-
-function declareWinner(){
-    black_background.style.display = "inline";
-    score.style.display = "block";
-    if(the_checker[1].color == "white")
-        score.innerHTML = "Black wins";
-    else
-        score.innerHTML = "Red wins";
-}
-
-
-function getDimension (){
-    contor ++;
-    windowHeight = window.innerHeight
-        || document.documentElement.clientHeight
-        || document.body.clientHeight;
-    windowWidth =  window.innerWidth
-        || document.documentElement.clientWidth
-        || document.body.clientWidth;
-}
-
-
-
-
-document.getElementsByTagName("BODY")[0].onresize = function(){
-
-    getDimension();
-    var cpy_bigScreen = bigScreen ;
-
-    if(windowWidth < 650){
-        moveLength = 50;
-        moveDeviation = 6;
-        if(bigScreen == 1) bigScreen = -1;
-    }
-    if(windowWidth > 650){
-        moveLength = 80;
-        moveDeviation = 10;
-        if(bigScreen == -1) bigScreen = 1;
-    }
-
-    if(bigScreen !=cpy_bigScreen){
-        for(var i = 1; i <= 12; i++){
-            b_checker[i].setCoord(0,0);
-            w_checker[i].setCoord(0,0);
-        }
-    }
-}
-
-
-
-
+    });
+};
